requirements.txt
argparse
requests
termcolor

pyyaml